requirements.txt
streamlit==1.13.0
pyyaml==6.0
numpy==1.23.4
pandas==1.5.1
splitwise==2.3.0
seaborn==0.12.1
streamlit-authenticator==0.2.1
